PSUSensor: Fix error for decimal part of scale factor
Customized scale factor declaration from schema in entity-manager is
declared as type "number." If a decimal number is used in PSUSensor,
it loses the decimal part when transforming the value from variant to
unsigned int, resulting in incorrect calculations.

Resolve this by using a variant to double and changing the scale factor
to double.

Tested:
Verified PSUSensor readings with customized decimal scale factor by
dbus introspect.

+++ src/psu/PSUSensor.cpp
@@ -52,7 +52,7 @@ PSUSensor::PSUSensor(
     boost::asio::io_context& io, const std::string& sensorName,
     std::vector<thresholds::Threshold>&& thresholdsIn,
     const std::string& sensorConfiguration, const PowerState& powerState,
-    const std::string_view sensorUnits, unsigned int factor, double max,
+    const std::string_view sensorUnits, double factor, double max,
     double min, double offset, const std::string& label, size_t tSize,
     double pollRate, const std::shared_ptr<I2CDevice>& i2cDevice) :
     Sensor(escapeName(sensorName), std::move(thresholdsIn), sensorConfiguration,
